+++ HISTORY.md
@@ -5,6 +5,8 @@
   through parallelization and avoiding logging. Avoid memory errors in segmentation.
 - UMI: upload prepared UMI bam file (pre-consensus) to final output directory
 - Add support for bbmap as an aligner
+- RNA-seq variant calling: parallelize GATK HaplotypeCaller over regions to
+  avoid memory and timeout issues.
 - CWL: adjust defaults for align_split_size and nomap_split_targets to match
   different parallelization and overhead for these runs
 - CWL: support for Cromwell runner